Fix world size in world select

+++ include/fs.h
@@ -22,7 +22,7 @@ extern "C"
     void fsDeleteFile(const char *name);
     bool fsFileExists(const char *name);
     char *fsReadFile(const char *name);
-    int fsGetFileSize(const char *name); // in bytes
+    long fsGetFileSize(const char *name); // in bytes
     char *fsHumanreadFileSize(double size);
 
 #ifdef __cplusplus
